feat: 集成 interview-reviewer 面试复盘 Skill（闭环最后一环）
- Skill（用户设计稿升内置，references 三规则表随包）
- migration 0005（no-transaction 重建 ai_runs）：run_type 增加
  interview_review；artifacts FK 保持
- Run 装配：勾选材料复制进 inputs/materials/ + rounds.json 轮次
 背景 + meta.json（公司/岗位/轮次）+ 上游 jd-analysis/
  tailoring-report 存在即带上（Skill 自判对账）；references 随 Run
- 提示词：无人值守一次性直出/降级产出；产物校验
  interview-review.json + .md + follow-up-email.md（transcript
  可选，存在即入库）；产物落 interviews/reviews/，信息型不推进状态
- 面试 Tab：材料勾选 + 启用「开始复盘」+ 复盘产物列表与查看；
  AI 记录类型标签补面试复盘
- specta BigInt 限制：size_bytes 用 i32

+++ apps/desktop/src-tauri/src/application/job_service.rs
@@ -813,7 +813,7 @@ impl JobService {
         .bind(kind)
         .bind(&raw_name)
         .bind(&rel)
-        .bind(meta.len() as i64)
+        .bind(meta.len() as i32)
         .bind(&now)
         .execute(&*self.db)
         .await?;
